requirements.txt
flask==3.0.3
youtube-transcript-api==1.2.4
google-generativeai==0.7.2
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
